Implement core upcall bridge
